Compliance with license: "unchanged tarball"
License only allows redistribution of all game data, (including bits which
otherwise are not redistributable) if the only change done to the tarball is
recompression "using different archival formats suitable for your OS
(zip/tgz/deb/dmg)"

Because of this, the redeclipse-data binary package now includes everything
from the original "tarball" (svn repo, as is the case here).

The redeclipse and redeclipse-server binary packages gets by under the
license of the source (Zlib) and enet (Expat). I Created a new icon (GPLv3)
since converting any of the redeclipse icon bases into an xpm is disallowed
as per the license.