Remove unmatched copyright fonts/* record, removed in repack
Update the copyright generation script to also filter out the fonts/*
content license records, since the contents are removed from upstream
tarball in the get-orig-source repack.

The Akashi font is part of some images/logos, but this should be a
document as far as OFL is concerned, and thus need not be licensed under
the OFL.

The Play font raster grids is rendered from ttf source in the fonts-play
package, and is not present at all in the redeclipse-data source
package.